Use ztimer_now_t instead of time_t
ztimer_now_t is 4 bytes, while time_t is 8 bytes, and printf doesn't
support %llu in this platform

+++ apps/wsn-main/main.cpp
@@ -15,7 +15,7 @@ int main(void)
     // Boot
     wsn_boot();
     LOG_INFO("app=wsn-main board=%s mcu=%s", RIOT_BOARD, RIOT_MCU);
-    LOG_INFO("basetime=%lld", wsn_time_basetime());
+    LOG_INFO("basetime=%ld", (long) wsn_time_basetime());
 
     // Start thread that handles incoming packets
     kernel_pid_t pid = thread_recv_start();
